+++ shared/filesystem.go
@@ -4,6 +4,6 @@ type Filename string
 type FileData []byte
 type FileInfo struct {
 	Creator         string
-	NumberOfRecords uint32
+	NumberOfRecords uint16
 	Data            FileData
 }
